+++ apps/api/src/app/user/user.module.ts
@@ -1,7 +1,9 @@
 import { OrderModule } from '@ghostfolio/api/app/order/order.module';
 import { SubscriptionModule } from '@ghostfolio/api/app/subscription/subscription.module';
+import { RedactValuesInResponseModule } from '@ghostfolio/api/interceptors/redact-values-in-response/redact-values-in-response.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { I18nModule } from '@ghostfolio/api/services/i18n/i18n.module';
+import { ImpersonationModule } from '@ghostfolio/api/services/impersonation/impersonation.module';
 import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 import { PropertyModule } from '@ghostfolio/api/services/property/property.module';
 import { TagModule } from '@ghostfolio/api/services/tag/tag.module';
@@ -18,6 +20,7 @@ import { UserService } from './user.service';
   imports: [
     ConfigurationModule,
     I18nModule,
+    ImpersonationModule,
     JwtModule.register({
       secret: process.env.JWT_SECRET_KEY,
       signOptions: { expiresIn: '30 days' }
@@ -25,6 +28,7 @@ import { UserService } from './user.service';
     OrderModule,
     PrismaModule,
     PropertyModule,
+    RedactValuesInResponseModule,
     SubscriptionModule,
     TagModule
   ],