+++ apps/client/src/app/components/account-detail-dialog/account-detail-dialog.component.ts
@@ -80,6 +80,7 @@ import { AccountDetailDialogParams } from './interfaces/interfaces';
 export class GfAccountDetailDialogComponent implements OnDestroy, OnInit {
   public accountBalances: AccountBalancesResponse['balances'];
   public activities: OrderWithAccount[];
+  public activitiesCount: number;
   public balance: number;
   public balancePrecision = 2;
   public currency: string;
@@ -100,7 +101,6 @@ export class GfAccountDetailDialogComponent implements OnDestroy, OnInit {
   public sortColumn = 'date';
   public sortDirection: SortDirection = 'desc';
   public totalItems: number;
-  public transactionCount: number;
   public user: User;
   public valueInBaseCurrency: number;
 
@@ -215,16 +215,17 @@ export class GfAccountDetailDialogComponent implements OnDestroy, OnInit {
       .pipe(takeUntil(this.unsubscribeSubject))
       .subscribe(
         ({
+          activitiesCount,
           balance,
           currency,
           dividendInBaseCurrency,
           interestInBaseCurrency,
           name,
           platform,
-          transactionCount,
           value,
           valueInBaseCurrency
         }) => {
+          this.activitiesCount = activitiesCount;
           this.balance = balance;
 
           if (
@@ -270,7 +271,6 @@ export class GfAccountDetailDialogComponent implements OnDestroy, OnInit {
 
           this.name = name;
           this.platformName = platform?.name ?? '-';
-          this.transactionCount = transactionCount;
           this.valueInBaseCurrency = valueInBaseCurrency;
 
           this.changeDetectorRef.markForCheck();