+++ CHANGELOG.md
@@ -9,7 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed the `transactionCount` in the portfolio calculator and service
+- Refreshed the cryptocurrencies list
+
+### Fixed
+
+- Fixed the accounts of the assistant for the impersonation mode
+- Fixed the tags of the assistant for the impersonation mode
+
+## 2.236.0 - 2026-02-05
+
+### Changed
+
+- Removed the deprecated `transactionCount` in the endpoint `GET api/v1/admin`
 - Upgraded `stripe` from version `20.1.0` to `20.3.0`
+
+### Fixed
+
+- Fixed an exception when fetching the top holdings for ETF and mutual fund assets from _Yahoo Finance_
 
 ## 2.235.0 - 2026-02-03